% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_env.R
\name{clear_env}
\alias{clear_env}
\title{Functions to clear environment}
\usage{
clear_env(env = .GlobalEnv)
}
\arguments{
\item{env}{the enironmen to clear}
}
\description{
Remove all objects from global environment default is global env
}
