% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_env.R
\docType{data}
\name{global_env}
\alias{global_env}
\title{A an example environment}
\format{
A environment with 7 objects
\describe{
\item{shop}{a dataframe that mimics a shopping list}
\item{todo}{a list that mimics a to do list. There is a list within it
called homework as well}
\item{quantity}{a numeric}
\item{item}{a character}
\item{availabe}{a logical}
\item{form}{a formula}
\item{sym}{a symbol}
}
}
\usage{
global_env
}
\description{
This envivonment is supposed to mimic a global environemnt for examples
and testing so that we do not screwup current .GlobalEnv
}
\keyword{datasets}
