% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{Peek at Envs or Lists}
\usage{
peek(x = .GlobalEnv)
}
\arguments{
\item{x}{environment or list that we would like to peak. Default is the
global environment}
}
\description{
Outputs a dataframe with name, class, length, dim for easy peeking at
workspace using the console
}
